rework out to pages
