+++ package.json
@@ -61,6 +61,7 @@
     "zod": "^3.24.1"
   },
   "devDependencies": {
+    "@types/leaflet": "^1.9.19",
     "@types/node": "^22",
     "@types/react": "^19",
     "@types/react-dom": "^19",
